#319: extract PeerTransferEngine out of PeerTransferComponent
Move PeerTransferState + Direction to domain package (transfer).
Create PeerTransferEngine — plain Kotlin, no Decompose/Compose deps —
owning all per-peer state-machine internals (inbound event collector,
outbound job, cancel/retry/retryFile logic, mapProgress/mapCompleted).
Shrink PeerTransferComponent to a delegating shell: bridges
engine.state:StateFlow to Value<PeerTransferState> via onEach/launchIn;
onCardClick handles picker vs pending logic; every other method is a
one-liner delegation.

Move all state-machine transition tests from PeerTransferComponentTest
to PeerTransferEngineTest (plain Kotlin, no ComponentContext needed).
Slim PeerTransferComponentTest covers only the bridge and onCardClick.
Update PeerListComponentTest, RootComponentTest, and
TransferDetailsComponentTest construction sites to construct an engine.
