--- composeApp/src/commonMain/kotlin/com/tubetoast/tether/presentation/transfer/Direction.kt
@@ -1,3 +1,0 @@
-package com.tubetoast.tether.presentation.transfer
-
-enum class Direction { Outbound, Inbound }
