+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/transfer/Direction.kt
@@ -1,0 +1,3 @@
+package com.tubetoast.tether.transfer
+
+enum class Direction { Outbound, Inbound }
