+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/transfer/PeerTransferState.kt
@@ -1,9 +1,4 @@
-package com.tubetoast.tether.presentation.transfer
-
-import com.tubetoast.tether.transfer.PartialOutcome
-import com.tubetoast.tether.transfer.PeerIdentity
-import com.tubetoast.tether.transfer.PerFileStatus
-import com.tubetoast.tether.transfer.TransferErrorReason
+package com.tubetoast.tether.transfer
 
 sealed interface PeerTransferState {
     val peer: PeerIdentity
